requirements.txt
numpy>=1.12.1
tensorflow-gpu>=1.1.0rc1
six>=1.10.0
Flask>=0.12
typing>=3.6.1